Add Lakeflow template (#2959)
## Changes

Add Lakeflow template based on the new Pipeline folder structure and
leveraging the new `glob` and `root_path` properties.

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->