testserver: Fix pipelines update to set default for storage (#2960)
## Why 
This is the default that is set by backend if field is missing in
Create/Update request. We already have Create covered, this applies it
to Update.

Discovered while working on terraform removal
https://github.com/databricks/cli/pull/2926
Terraform reads the pipeline from the backend and then adds those
defaults to future update requests while the new deployment backend does
not read anything from the backend and those fields remain empty in
update requests. The end result (as measured by "pipelines get $id") is,
however, the same.