Deduplicate TerraformResourceName (#2995)
## Changes

The `TerraformResourceName` field was added to the description struct in
#1928. This PR removes the duplication between that field and the
`TerraformResourceName` function on all resources. It also uses this
value when interpolating Terraform resource references instead of
keeping another hardcoded value around. The value for the model serving
endpoint is updated to reflect the right value (it wasn't used before).


## Tests

Existing tests pass.