Tolerate URLs copied from the browser with `databricks auth login` (#3001)
## Changes
It's a common flow for a user to copy the URL of their workspace from
the browser to login with the CLI. All that we need to successfully
login is the hostname, but the URL contains much more information, such
as the path of the page they were on or their workspace ID as a query
parameter.

Today, if a user tries to use this URL verbatim with `databricks auth
login`, they get an error like `Error: fetching oauth config: fetching
OAuth endpoints: databricks OAuth is not supported for this host`.

This PR ensures that the CLI uses the same normalization as the Go SDK
config on the provided hostname before starting the OAuth login flow,
resolving this error.

## Why
This reduces friction for those logging in with the CLI.

## Tests
Unit tests to verify that such URLs are tolerated.