Rename jsontag to structtag; support bundle tags (#2976)
## Why
We use bundle tags in our structs to table internal / readonly or
deprecated fields. This is a general package that allows us to inspect
those tags during a `structwalk.Walk`. Support for `structwalk.Walk`
will be added in a followup PR.

JSON tags can also be accessed via the same package.

## Tests
Unit test.