Fix dashboard generation for already imported dashboard (#3016)
## Changes
We started inlining the dashboard definition in
https://github.com/databricks/cli/pull/2738 by reading `file_path` and
setting it in `serialized_dashboard`.

This broke the `bundle generate` CUJ for updating an already imported
dashboard when working with DABs. This PR fixes that by keeping the
`file_path` around in the configuration tree as metadata and only
dropping it when we generate the TF configuration.

This PR also adds regression tests to ensure this keeps working in the
future.

## Why
Fixes
https://github.com/databricks/cli/issues/2998#issuecomment-2950723565.

## Tests
Acceptance tests.