Enable yaml formatter; fix formatting in default-python & dbt-sql (#3026)
## Changes
- Update 'make fmt' to format yaml with
https://github.com/google/yamlfmt/
- Templates dbt-sql and default-python updated to have consistent
indentation according to formatter.

## Why
For templates with a lot of options (default-python) it's easy to
introduce formatting issues - too many empty lines or not enough in the
output. I'd like to use formatter also on non-materialized templates in
default-python/combinations to ensure that resulting yamls are formatted
well for any possible combination of options. Related:
https://github.com/databricks/cli/pull/3004