Translate dashboards' paths before their serialized contents are read (#3006)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

#2738 changed reading the dashboard contents before the dashboard path
was normalized. This change fixes that defect, and also fixes use cases
of reading dashboards contents inside a bundle, which defines sync root
that is different from the bundle root

## Tests
<!-- How have you tested the changes? -->
Existing and new acceptance tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->