Add DELETE endpoint for lakeview dashboards in test server (#3045)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. Add DELETE endpoint for lakeview dashboards in test server
1. Remove the mocks for DELETE calls in acceptance tests for dashboards
deployment

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This is the first in the series of changes for using testserver instead
of mocks in acceptance tests for dashboards deployment

## Related PRs:
#3047

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->