Remove all server mocks from bundle/deploy/dashboard acceptance tests (#3048)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This is the last in the series of changes for using testserver instead
of mocks in acceptance tests for dashboards deployment

## Related PRs
- #3045
- #3047

## Tests
<!-- How have you tested the changes? -->
Existing tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->