Narrow scope of telemetry acceptance tests (#3053)
## Changes
This PR narrows the telemetry acceptance tests to on assert on the
interesting part of the telemetry request payload instead of asserting
on the entire payload.

## Why
Having more targeted tests will result in a smaller difference when
adding new metrics that we want to track in telemetry. Today, any new
fields added to telemetry end up touching multiple tests. This PR will
prevent that moving forward.