Refactor: extract translatePathsDashboards into a separate mutator (#3031)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
This change was suggested in #3006
[comment](https://github.com/databricks/cli/pull/3006#discussion_r2139375127)

This simplifies the translatePaths mutator and adds explicit naming for
the translatePathsDashboards, which is visible in debug logs.

## Tests
<!-- How have you tested the changes? -->
Existing acceptance tests for dashboards deployment

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->