Convert TestDeployBundleWithApp to an acceptance test (#3066)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
Convert TestDeployBundleWithApp to an acceptance test. Note, that this
new acceptance test only has a slow version of it. The initial (fast)
part of the original test is already covered by existing
`acceptance/bundle/apps` tests

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
One change in the series of changes for converting integration tests
into acceptance tests.
This will allow for easier testing of various backing solutions for
bundle deployment

## Testing
- Running the new test from local machine
- Successful run of this PR:
[go/deco-tests/15710009578](http://go/deco-tests/15710009578) (with
Cloud=true, which triggered the new test to run against all stages)


<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->