Remove TestBindSchemaToExistingSchema integration test (#3073)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

TestBindSchemaToExistingSchema is equivalent to an existing acceptance
test in acceptance/bundle/deployment/bind/schema introduced in #2441

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->