Allow users to override the Terraform version to use (#3069)
## Changes

Previously, the CLI would use the Terraform binary at
`DATABRICKS_TF_EXEC_PATH` only if the `DATABRICKS_TF_VERSION`
environment variable was set and matched the _built-in default version_.
It was built such that we could ship all binaries with the VS Code
extension to avoid re-downloading (see #1294).

This change expands the scope of these variables and makes them
composable. Now, you can:
1. Configure `DATABRICKS_TF_VERSION` to override the Terraform version
to use
2. Configure `DATABRICKS_TF_EXEC_PATH` to configure the path to an
existing Terraform binary

If 1 is used alone, the CLI will download the specified version. If 2 is
used alone, the CLI will execute the specified binary to confirm that it
matches the _built-in default version_. If both are used, the CLI will
execute the specified binary to confirm that it matches the specified
version.

## Why

To allow users to use newer version of the Terraform CLI.

## Tests

* Unit tests pass
* Integration tests with the latest Terraform version (#3040)