Make FromSlice and ToSlice methods private to the `deploy` package (#3089)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

Keeping methods private to the package makes the code easier to
comprehend - private methods are used only inside the package where they
are defined

## Tests
<!-- How have you tested the changes? -->
Existing tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->