Convert TestAbortBind to an acceptance test (#3116)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
One change in the series of changes for converting integration tests
into acceptance tests.
This will allow for easier testing of various backing solutions for
bundle deployment

This is a follow-up of the closed not-merged PR #3111