Fix failing `serialized_dashboard` acceptance tests (#3119)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. change the `dashboard/detect-change` acceptance test, so that it only
checks for a single field in the `serialized_dashboard` output
2. change the `bind/dashboard` acceptance tests so that it normalizes
the contents of the `serialized_dashboard` field before printing to
output

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
Due to changes in the Lakeview API these tests started failing in our
staging environments

## Tests
<!-- How have you tested the changes? -->
Ran the changed tests from local machine against AWS and GCP targets

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->