Changes because of always defining request messages (#3098)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
Changes because of always defining request messages. There is a bug in
the Code Generator that sets the request to nil if it has no fields. As
we fix the generator, it changes the generated code. To keep the
generated code the same, we need to add exceptions for already existing
requests.
IsLegacyEmptyRequest returns true if the request has no fields and, for
the above reason, is not rendered in the SDK.

This PR also removes TODO from the template, which adds noise to the
diff. Also, we don't need it to remind us to implement short flags.

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
We made a few changes in the Code generator that ensure that the request
can never be nil. However, we cannot add the request to the existing
interface. So, adding an exception for legacy commands.

## Tests
<!-- How have you tested the changes? -->
Existing CI. Manually verified this change, only changed a few comments.

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->