summary/run/open: pull App and SecretScope's name into ID field, not Name (#3122)
## Changes
When doing Terraform.Load (bundle summary/run/open), do not override
app.name and secret_scope.name fields that come from the config with the
one stored in Terraform state. Instead store state value in the "id"
field.

## Why
Conceptually, we separate configuration field (Name) and stored state
field (ID, which is Name from last deployment). This means as a user I
can see both. Today, configuration's Name is lost in "bundle summary".
However, in "bundle validate" Name would be equal to the one from the
config, which is confusing -- potentially they could be different.

Internally, this makes TerraformToBundle simpler - it only updates ID
field and ModifiedStatus (both are always computed and don't come from
the user). This means configuration is preserved and also we can make
TerraformToBundle generic so that it does not need to have per-resource
code (which becomes 2x when we add direct deployment method support).

## Tests
Existing tests.