acc: Fixed test output for dashboard and jobs URLs (#3144)
## Changes
Fixed test output for dashboard and jobs URLs

## Why
Integration tests are failing with 
```
        -  "url": "[DATABRICKS_URL]/dashboardsv3/[ALPHANUMID]/published?o=[NUMID]",
        +  "url": "[DATABRICKS_URL]/dashboardsv3/[ALPHANUMID]/published",
```

and

```
        -  "url": "[DATABRICKS_URL]/jobs/[NUMID]?o=[NUMID]"
        +  "url": "[DATABRICKS_URL]/jobs/[NUMID]"
```

We only started to track this output on Cloud runs in this PR
https://github.com/databricks/cli/pull/3130/files and it was failing
from the start.

So it's not a change in our code or backend which lead to the diff.
Likely `?o=` is omitted on the cloud runs because orgId is already
included in workspace URL and hence `?o=` not added
https://github.com/databricks/cli/blob/cf1495663458afc1b12649505ffb00e62f794974/bundle/config/mutator/initialize_urls.go#L47-L58


## Tests
Covered with existing tests

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->