Cleanup scripts for windows in ShellExec (#2902)
## Why
The `cmd.exe` command executor we use to execute scripts in windows
writes the script first to disk before executing it. This is necessary
because `cmd.exe` does not support inlining the script contents as an
argument.

Since ShellExecv needs to exit the current CLI process within the
confines of the `execv` method to mirror the behaviour of the UNIX
version of `execv`, this PR adds a callback that cleans up the temporary
script before exiting the `execv` function call.

## Tests
Unit tests which verify the script is indeed being cleaned up.