Generalize terraform to bundle conversion, prepare for direct deployment (#3134)
## Changes
- Introduce ExportedResourcesMap type that is deployment-method neutral.
Modify terraform.ParseResourcesState to return this type instead of
struct matching terraform state format.
- Modify terraform.TerraformToBundle() to rely on this new format.
- Make terraform.TerraformToBundle() handle all resources automatically
and operate on dynamic bundle structure.
- Remove ModifedStatusUpdated constant and usage in apps.
- Fix Secret scopes resource to properly set modified_status. Previously
it was always "created".

## Why
ExportedResourcesMap allows implementing terraform.Load() and thus
bundle summary/run/open with direct deployment method
(https://github.com/databricks/cli/pull/2926) as it is not tied to
terraform resource names or internal state structure.

When implementing new resource, we no longer need to add an entry in
TerraformToBundle, this both saves time and avoids bugs.

ModifiedStatusUpdated is removed because it's not implemented. Only app
resource ever sets it and it sets it always, not based on the actual
situation.

## Tests
summary/modified_status is expanded to handle both cases where
resources: block is present and not present (regression during
development of this PR).