Add a readonly function to walk a dynamic tree (#3084)
## Changes
This PR adds a read-only version of dyn.Walk, which can be used to walk
the bundle configuration tree faster than the normal walk method.

Benchmarks show it is 4x faster than the normal walk method.

## Why
Many walk use cases are read-only, so having such a method is useful.
It'll be used to perform required and enum validation in DABs.

## Tests
New unit tests.