acc: Use more specific regex in default-python/combinations (#3204)
This test uses regex to verify that if include_python is set to "yes"
then we have a line like this:

Uploading
.databricks/bundle/dev/patched_wheels/python_artifact_x[UNIQUE_NAME]/x[UNIQUE_NAME]-0.0.1+[UNIX_TIME_NANOS]-py3-none-any.whl...

and if include_python is set to "no" then we don't have a line like that
in the output.

However, we also always print a line like this:

Uploading bundle files to
/Workspace/Users/[USERNAME]/.bundle/X[UNIQUE_NAME]/dev/files...

Since UNIQUE_NAME is random string, it sometimes contains "whl", causing
regex to match where it should not.