Rename CacheDir to LocalStateDir (#3194)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
1. Rename `CacheDir` interface to `LocalStateDir`
2. Rename tests that mention `CacheDir` so that they mention
`LocalStateDir` now

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->
`CacheDir` is a term that is not semantically correct since it contains
not only cache, but also the state of the bundle.

## Tests
<!-- How have you tested the changes? -->
Existing tests are passing

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->