Normalize clusters in jobs according to Terraform provider logic (#3198)
## Changes

Apply changes in resource normalization that are applied by TF and cause
invalid configuration to be fixed and accepted by the backend.

For example, when InstancePoolId is set, clear NodeTypeId field.


## Why

For direct deployment we do all resource transformations in initialize
phase rather than before deployment.

This makes no effect on the resource - the APIs are called with the same
payload because this transformations are idempotent when applied for the
second time.

However,
- this makes it visible to user (via bundle validate -o json) what is
the final configuration that is being sent to the backend
- makes the same configuration work equally in terraform and direct
backend even through direct backend does not transform resources in any
way.

Same reasoning as in https://github.com/databricks/cli/pull/2767

## Tests
A few tests that were failing on direct backend are now enabled.

---------

Co-authored-by: Pieter Noordhuis <pieter.noordhuis@databricks.com>