Upgrade Go SDK to 0.75.0 (#3256)
## Changes
Upgrade Go SDK to 0.75.0

The biggest difference is that `genkit` does not infer field description
from the custom types description anymore (for example fields like
`data_security_mode`), thus we do it for annotations ourselves.

The result is that we are now missing descriptions for some of the enum
flags.

## Tests
Existing tests pass

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->