Configure `__apply_policy_default_values_allow_list` for task and job clusters (#3255)
## Changes

Automatically reapply cluster policy defaults for task and job clusters
when updating existing jobs.

Note: if you relied on task and job cluster settings that originate from
a cluster policy _not_ changing on deploy, even though the cluster
policy defaults they were drawn from did change, then you'll need to
explicitly include those values in your bundle configuration from now
on.

## Why

The latest version of the Terraform provider comes with a feature
specifically for re-applying cluster policy defaults for task and job
clusters when updating a job
(https://github.com/databricks/terraform-provider-databricks/pull/4834).
Before this change, it was not possible to reapply cluster policy
defaults for an existing job, as Terraform keeps server-side default
values in its state. Users had to destroy and recreate jobs for new
defaults to take effect.

With the Terraform dependency in DABs, the same happened for DABs users,
and cluster policy defaults were effectively "snapshotted" on the first
deploy, and never updated thereafter.

This is contrary to the behavior we expect to see. Unless a field is
explicitly configured in the bundle configuration, cluster policy
defaults should be applied on every deploy, regardless of the job being
created or updated.

This change passes the list of fields explicitly specified in the bundle
configuration to Terraform via the
`__allow_policy_default_values_allow_list` property introduced in the
linked PR.

## Tests

* Unit tests pass.
* Manually confirmed that when the cluster policy for a job cluster is
changed, the fields that aren't explicitly specified in the bundle
configuration draw their default values from the new cluster policy.

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->