Change execv to use current directory as a working directory if no directory is provided (#3267)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

This is a follow-up for the comment in the previous PR #3128 :
https://github.com/databricks/cli/pull/3128#discussion_r2205016022

## Tests
<!-- How have you tested the changes? -->
- Existing acceptance test on cmd/psql is still passing even though the
production code is not specifying a directory
- Manually checked that when executed with `pwd` the process prints
current working directory on Mac and on Windows

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->