acc: Extract token calculation into common block (#3314)
## Changes
Use the same code for calculating token to use in tests.

## Why
There is no need to have different logic for startProxyServer and
startLocalServer and there is no reason to make those functions
responsible for the token.

I'd like to use token to encode other info, such as user name or service
principal name, this helps with that.