Show "DEFAULT" as the default profile for `databricks auth login` (#3252)
## Changes
Show `"DEFAULT"` as the default profile for the `databricks auth login`
command. This change uses the same UX as seen in the `databricks bundle
init` command, i.e. it shows the default value between square backets
rather than in the input field.

## Why
This change aims to make `databricks auth login` more self-documenting.
I want to reference it from templates in the future.