direct: refactor deploy/plan interaction
"bundle deploy" no longer calculates and classifies the diff, it relies
on action types calculated during "bundle plan". (Note, this will need to
be further developed once we handle ${resources} references - we might want to
recalculate the diff in some cases.).

We no longer have UpdateUpdatesID setting on the resource. Instead,
resources can implement optional method DoUpdateWithID that can return an
updated id as part of the update. Currently, only volume uses that.

The original DoUpdate method no longer allows to return ID. By
convention, if possible, it also checks that ID is not changed by the
backend.

There is new ActionType: "update_with_id". This is what triggers
DoUpdateWithID() call. However, when stringified, this is presented as
"update" so that it's compatible with terraform output in "bundle plan".

The main motivation is to have information at plan-time on whether we're
going to update ID or not. If not, we can resolve ${resources.*.*.id}
right away and have more parallelism at deploy time.