pipelines init: make dev and prod different when specifying schema (#3303)
## Changes
To cli-pipelines template, append .dev and .prod to distinguish between
schemas in prod and dev, avoiding overwriting of data when initializing
pipelines template with a custom schema in prompt or JSON.

Furthermore, the prompt to the user for a schema explains that a
specific schema will have .dev and .prod appended to it.
In the README, the full name of the pipeline is now noted, along with
where to find the pipeline.

## Tests
SQL test, which uses the shared schema, updated to print an example
template with the new schema.