Add `--omit-repl` flag to the gh_report tool (#3413)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

Skipping the lines that print out available replacements when reading
out failed tests log significantly reduces noise and cognitive load for
the reader.

## Tests
<!-- How have you tested the changes? -->
Manually ran `gh_report` with the flag and without the flag and
confirmed that both outputs still work

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->