Resolve $resources refs before deployment if set by user (#3423)
## Changes
Resolve ${resources} references internally, before the deployment.

This is v2 of https://github.com/databricks/cli/pull/3370 which was
reverted in https://github.com/databricks/cli/pull/3421 because it
interacted poorly with other resources transformations that are done by
DABs. By moving this resolution later we ensure that all transformation
are visible when we do reference replacement.

## Why
Currently we resolve these via terraform, but that's not going to work
with direct deployment. To minimize the changes between TF and direct
and surface possible issues with this approach, we do this for both
backends.

---------

Co-authored-by: shreyas-goenka <88374338+shreyas-goenka@users.noreply.github.com>