Fix "cache: token not found" for auth token command (#3447)
## Changes

Update the "auth token" command to return the same error as CLI
v0.263.0.

## Why

Older versions of Databricks SDKs call this command to retrieve an OAuth
token. They perform a substring match on the returned error to determine
if the CLI experienced a real error or if OAuth is not an applicable
authentication mechanism in the user's context. By changing the error
message (see databricks/databricks-sdk-go#1250), these older SDKs no
longer fall through but return an error `cache: no token found` when
checking if they can use OAuth to authenticate.

## Tests

Adds an acceptance test. It passes with the previous version of the CLI
and passes with this commit.