Add downloaders for uv, go and jq (#3444)
## Why
We'll package the CLI repo and binaries for uv, go and jq into a tar.gz
archive before uploading it to a serverless cluster for execution. This
PR adds the downloader for those tools. A follow-up PR will be sent for
actually building the archives.

## Tests
Unit tests. The tests have a "dbr_only" build tag added, so they'll not
run during normal execution. I plan to run them only from the CI
integration run job for DBR.