Show a warning message if a resource is not in GroupToTerraformName (#3463)
## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

When introducing a new resource to the system it does not get a mention
in the plan automatically. Instead, developer must know that the new
resource has to be added to the GroupToTerraformName map for it. The
warning message is a good indicator for the developer of the new
resource that an extra step is needed for the resource support. I opted
for the warning message instead of panicking because apparently there
are resources that do not need to be in the plan, but there is no
comprehensive list of such resources provided and I am not sure that the
test suite covers every such resource, it is not a great experience for
the bundle operation to be blocked just because the resource is in
neither of the lists.

## Tests
<!-- How have you tested the changes? -->
Existing tests are green

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->