Add archive build for DBR testing (#3452)
## Changes
This PR adds an internal program that can build a CLI test archive that
contains the entire repo and binaries like uv, go, and jq to run
acceptance tests.

This PR builds on top of https://github.com/databricks/cli/pull/3444

## Why
A single deployable binary makes testing on DBR very straightforward.
All you then need to do is decompile the binary and run tests using a
test runner of your choice.

For now, the tests will be run using the go binary, but this could be
extended to run them using gotestsum.

## Tests
A new unit test. This has a build tag so that we can limit its execution
to only a DBR CI environment.