Add -workspace-tmp-dir option to acceptance tests (#3498)
## Changes
The—workspace-tmp-dir option makes the acceptance testing framework run
the test using the workspace file system instead of the local file
mount.

## Why
This option is necessary to simulate DABs in the workspace, since it
runs the Databricks CLI on the workspace file system.

## Tests
Manually. I've been using this function to run a couple of test runs in
my other PR: https://github.com/databricks/cli/pull/3453