acc: Use fixed ports instead of dynamically allocated ones for run-local tests (#3490)
## Changes
Use fixed ports instead of dynamically allocated ones for run-local
tests

## Why
After researching and trying to stabilize the tests, I concluded that
dynamic port allocation is likely the root cause of its flakiness. The
hypothesis is that `allocate_ports.py` calls socket.close() after
finding a free port. When the socket is closed, the port it was using is
released back to the OS. However, the port does not become instantly
available for another application to use. It typically enters a
TIME_WAIT state to handle any stray packets.
It could be that the "connection refused" error, which happens when the
curl command tries to connect after the Python script has closed the
socket, but before the app has successfully bound to the port and is
ready to accept connections

## Tests
After making ports fixed in the script, all tests for all runners
succeeded (tried 5 times)

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->